requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
gtts==2.5.1
boto3==1.35.0 